refactor MessageManagers
