+++ newcell/managers/imu_message_manager.py
@@ -1,8 +1,6 @@
 import pandas
-import pytz
-from datetime import datetime, timedelta
-from timezonefinder import TimezoneFinder
 
+from newcell.managers.message_manager import MessageManager
 from newcell.messages.imu_message import ImuMessage
 
 
@@ -22,36 +20,10 @@ HEADER_IMU = [LABEL_DATETIME,
               LABEL_IMU_GX, LABEL_IMU_GY, LABEL_IMU_GZ,
               LABEL_IMU_MX, LABEL_IMU_MY, LABEL_IMU_MZ]
 
+@MessageManager.register
 class ImuMessageManager:
-
-    # TODO: (if needed) imu message validation
-    # length / value validities
-
     def __init__(self):
         self.messages = []
-
-    @staticmethod
-    def adjust_timezone(dataframe: pandas.DataFrame, tz) -> pandas.DataFrame:
-        start_at = dataframe.at[0, LABEL_DATETIME]
-        dataframe[LABEL_DATETIME] += tz.utcoffset(start_at)
-
-        return dataframe
-
-    @staticmethod
-    def adjust_timezone_by_coordinate(dataframe: pandas.DataFrame, lat: float, long: float) -> pandas.DataFrame:
-        tz = pytz.timezone(TimezoneFinder().timezone_at(lat=lat, lng=long))
-
-        return ImuMessageManager.adjust_timezone(dataframe, tz)
-
-    @staticmethod
-    def adjust_date(dataframe: pandas.DataFrame, date: datetime) -> pandas.DataFrame:
-        date_delta: timedelta = date.date() - dataframe.at[0, LABEL_DATETIME].date()
-
-        dataframe[LABEL_DATETIME] = dataframe[LABEL_DATETIME].map(
-            lambda dt: dt + date_delta,
-        )
-
-        return dataframe
 
     def add_message(self, payload: bytes) -> None:
         self.messages.append(ImuMessage.create(payload).export_row())
